requirements.txt
asgiref==3.10.0
attrs==25.4.0
autobahn==24.4.2
automat==25.4.16
certifi==2025.11.12
cffi==2.0.0
channels==4.3.1
charset-normalizer==3.4.4
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
django==4.2.25
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
filelock==3.19.1
fsspec==2025.10.0
huggingface-hub==0.36.0
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
jinja2==3.1.6
markupsafe==3.0.3
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
packaging==25.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyopenssl==25.3.0
pyyaml==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
sentencepiece==0.2.1
service-identity==24.2.0
sqlparse==0.5.3
sympy==1.14.0
tokenizers==0.22.1
tomli==2.3.0
torch==2.8.0
tqdm==4.67.1
transformers==4.57.1
twisted==25.5.0
txaio==23.6.1
typing-extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
zope-interface==8.0.1
